requirements.txt
attr==0.3.2
avalanche_lib==0.5.0
ConfigParser==6.0.1
contextlib2==21.6.0
cryptography==42.0.5
Cython==3.0.9
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.2
Jinja2==3.1.3
jnius==1.1.0
keyring==25.0.0
lockfile==0.12.2
matplotlib==3.7.5
mock==5.1.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.2.0
protobuf==5.26.0
pyOpenSSL==24.1.0
pytz==2024.1
railroad==0.5.0
scikit_learn==1.3.2
Sphinx==7.2.6
thread==1.1.1
torch==1.10.1+cu111
torchvision==0.11.2+cu111
tornado==6.4
trove_classifiers==2024.3.25
truststore==0.8.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
